+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 # Changelog
 
+## v2.0.0 2019-07-01
+
+- `namingConvention` config option is now the same as for `base-visitor` (`module#method`).
+
 ## v1.3.1 2019-07-01
 
-* Updated to graphql-codegen 1.3.1.
+- Updated to graphql-codegen 1.3.1.
 
 ## v1.2.0 2019-05-20
 
-* Initial release based on graphql-codegen 1.2.0.
+- Initial release based on graphql-codegen 1.2.0.